Bugfix: Require BigQuery queries to reference at least one table.
A table-free SELECT (`SELECT SESSION_USER()`, `@@project_id`) dry-runs
with empty referencedTables, was allowed for project-only scope, tracked
zero datasets, and executed with the owner's token -- returning
owner/session context no dataset probe can verify an observer against.
An empty tracked set admits any verifier without a single probe.
INFORMATION_SCHEMA reads that dry-run with empty refs rode the same
path.

now refuses empty referencedTables on every binding shape, with a
message pointing the agent at gadget code for table-free computations.
The scalar-function-with-a-table variant (SELECT SESSION_USER() FROM t
LIMIT 1) is unfixable at the referencedTables level and is documented as
part of the existing metadata-class residual.

BigQuerySessionImpl moves to its own module so Node vitest can construct
it directly (google.ts imports workerd-only modules and bundled assets);
the tests stub `cloudflare:workers` and `capnweb-validate`, the same
pattern as mcp-shared.
