+++ docs/observers.md
@@ -662,7 +662,7 @@ its resource types.
 | **google** | Google Calendar (selected calendar) | **B** | Require `writer` or `owner` access to the bound calendar, since `reader` access hides private-event details. Future: let the binding owner exclude private events so readers can collaborate. |
 | **google** | Google Calendar (`allVisible` availability) | **C** | In addition to the selected-calendar check, track foreign calendars whose free/busy data was successfully read and verify each observer can independently query their availability. |
 | **google** | Gmail Mailbox | **A** | Always throw. (Future: allow observers who independently have access, e.g. mailing-list members — explicitly out of scope now.) |
-| **google** | BigQuery | **C** | Track accessed datasets; verify the observer's IAM access to each. Dataset granularity for now (tables/columns later). The probe (`datasets.get`) is also metadata-level: it proves `bigquery.datasets.get` only — not `bigquery.tables.getData` on the tables queries actually read, nor row-level security, nor column policy tags — so an admitted observer may see query rows they could not query themselves. Known, accepted limitation; tightening direction is per-table tracking (the dry-run's `referencedTables`) plus a table-data probe, with row/column policies unprovable via any API regardless. |
+| **google** | BigQuery | **C** | Track accessed datasets; verify the observer's IAM access to each. Dataset granularity for now (tables/columns later). The probe (`datasets.get`) is also metadata-level: it proves `bigquery.datasets.get` only — not `bigquery.tables.getData` on the tables queries actually read, nor row-level security, nor column policy tags — so an admitted observer may see query rows they could not query themselves. Known, accepted limitation; tightening direction is per-table tracking (the dry-run's `referencedTables`) plus a table-data probe, with row/column policies unprovable via any API regardless. Queries whose dry run references no tables at all are refused outright — their access can't be attributed to any dataset, and they can return owner/session context (`SESSION_USER()`) no probe covers. |
 | **linear** | Team / Issue | **B** | Check the observer's workspace/team membership, honoring team privacy. |
 | **linear** | Workspace | **C** | Track accessed teams; verify the observer against each (reusing the Team B check). |
 | **notion** | Page / Database | **B** | Check the observer's Notion access to the bound page/database. |
